Add `core/util_rules/system_binaries.py` and relocate `archive.py` types like `UnzipBinary` to it (#14601)
## Motivation

Have one central location where all our system binaries are set up. We have two objectives with system binaries:

1. Ensure people use `BinaryPath` rather than hardcoding `/bin/tool`, which is necessary for working on diverse machines like NixOS https://github.com/pantsbuild/pants/issues/14492.
2. Limit the number of binaries we use. It has some performance hit + increases the requirements for Pants to work.

By centralizing _all_ of our system binaries code here, I think we will better achieve both goals. For example, https://github.com/pantsbuild/pants/issues/14492 proposes a new `[system-binaries]` subsystem (credit @benjyw for name), and this gives us a single place to make sure every binary is wired up to the subsytem.

## Followup
1. Move `BinaryPath` and `BashBinary` from `pants.engine.process` to this file. Also `PythonBinary` here.
      - We put these types in `process.py` because we didn't have a home at the time. Imo, they belong better in `core/util_rules` than `engine/`: these are not foundational types to the engine. They are conveniences, similar to `external-tool.py` helping to download tools.
      - This will be a breaking change. We can't re-export the types due to import cycles.
4. Add some new binaries like `Mkdir`.

[ci skip-rust]
